+++ src/components/Dialogs/RemoveBirdFromObservation.vue
