+++ src/components/Dialogs/RemovePlace.vue
